afxdp/cos/token_bucket: relocate inline tests to token_bucket_tests.rs
Same playbook (#1080-#1116, 33× successful today). 186-LOC inline test
block extracted. token_bucket.rs: 384 → 201 LOC.
cargo build/test 92/878.
